iter14 fix 8 refactoring
