+++ internal/app/handlers/pingdb.go
@@ -1,0 +1,26 @@
+package handlers
+
+import (
+	"context"
+	"database/sql"
+	"net/http"
+	"time"
+
+	"github.com/gin-gonic/gin"
+)
+
+func (handler *URLHandler) DBPingConn(c *gin.Context) {
+	db, err := sql.Open("pgx", handler.dbPath)
+	if err != nil {
+		http.Error(c.Writer, err.Error(), http.StatusInternalServerError)
+	}
+	defer db.Close()
+
+	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
+	defer cancel()
+	if err = db.PingContext(ctx); err != nil {
+		http.Error(c.Writer, err.Error(), http.StatusInternalServerError)
+	}
+
+	c.Writer.WriteHeader(http.StatusOK)
+}
